requirements.txt
accelerate==0.20.3
appdirs==1.4.4
attrs==23.1.0
azure-ai-formrecognizer==3.3.0b1
azure-common==1.1.28
azure-core==1.27.1
backoff==2.2.1
boilerpy3==1.0.6
canals==0.2.2
cattrs==23.1.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.4
dill==0.3.6
docopt==0.6.2
Events==0.4
exceptiongroup==1.1.1
farm-haystack==1.17.1
filelock==3.12.2
fsspec==2023.6.0
generalimport==0.3.1
huggingface-hub==0.15.1
idna==3.4
inflect==6.0.4
isodate==0.6.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
lit==16.0.6
MarkupSafe==2.1.3
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
msrest==0.7.1
networkx==3.1
nltk==3.8.1
num2words==0.5.12
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
packaging==23.1
pandas==2.0.2
Pillow==9.0.0
posthog==3.0.1
protobuf==3.20.2
psutil==5.9.5
pydantic==1.10.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
quantulum3==0.9.0
rank-bm25==0.2.2
regex==2023.6.3
requests==2.31.0
requests-cache==0.9.8
requests-oauthlib==1.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sseclient-py==1.7.2
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.1
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
url-normalize==1.4.3
urllib3==2.0.3
